nlp/requirements.txt
torch>=1.10.0
transformers>=4.18.0
datasets>=2.0.0
seqeval>=1.2.2
pandas>=1.3.0
numpy>=1.20.0
pyyaml>=6.0
tqdm>=4.62.0
